Fold old dynamic resource tests -> test_simplates
Since that's really what they were testing.
